+++ test/simple/test.sh
@@ -2,6 +2,8 @@
 set -euo pipefail
 
 cd "$(dirname $0)/app"
+source '../../test-helpers.sh'
+
 declare -r output='simple_0.1.0_all'
 
 finish() {
@@ -14,8 +16,8 @@ trap 'finish' EXIT
                   --no-delete-temp \
                   -- app.js lib/
 
-grep -q 'Package: simple' "$output/DEBIAN/control"
-grep -q 'Version: 0.1.0' "$output/DEBIAN/control"
+grep -q 'Package: simple' "$output/DEBIAN/control" || die 'Incorrect package name'
+grep -q 'Version: 0.1.0' "$output/DEBIAN/control"  || die 'Incorrect package version'
 
 dpkg -i "$output.deb"
 apt-get purge -y simple
